fixed file .read() with `.encode('utf-8'))` for open files mode="wb" and  `.decode('utf-8')` for mode="rb" in many places

+++ lib/jsontemplate.py
@@ -58,7 +58,7 @@ __all__ = [
     'TemplateSyntaxError', 'UndefinedVariable', 'CompileTemplate', 'FromString',
     'FromFile', 'Template', 'expand']
 
-import StringIO
+from io import StringIO
 import pprint
 import re
 
@@ -743,7 +743,7 @@ def SplitMeta(meta):
     if n % 2 == 1:
         raise ConfigurationError(
             '%r has an odd number of metacharacters' % meta)
-    return meta[:n / 2], meta[n / 2:]
+    return meta[:int(n/2)], meta[int(n/2):]
 
 
 _token_re_cache = {}
